Move ColorsManager to config module
